starting to test batch file processing. Very messy code

+++ main.py
@@ -32,6 +32,10 @@ import sys
 import optparse
 import logging
 import os
+import re 
+import time 
+import copy 
+import subprocess
 from os.path import basename, splitext
 from psat import read_psat, read_contingency
 from parselog import SimInLimits
@@ -49,35 +53,77 @@ logger = logging.getLogger(__name__)
 # Grem: regexp file deleter
 #------------------------------------------------------------------------------
 
-def grem(path, pattern):
-	pattern = re.compile(pattern)
-	for each in os.listdir(path):
-		if pattern.search(each):
-			name = os.path.join(path, each)
-			try: os.remove(name)
-			except:
-				grem(name, '')
-				os.rmdir(name)
+def grem(path, pattern, test=False):
+    pattern = re.compile(pattern)
+    for each in os.listdir(path):
+        if pattern.search(each):
+            name = os.path.join(path, each)
+            try:
+                if not test: os.remove(name)
+                logger.info("Grem removed " + name)
+            except:
+                if not test: 
+                    grem(name, '')
+                    os.rmdir(name)
+                logger.info("Grem removed dir " + name)
 
+def test_grem():
+    grem(".", ".*\.pyc", True)
+    grem(".", "rts_[1234567890]{2}\.txt", True)
+# test_grem()        
+        
 #------------------------------------------------------------------------------
 # 
 #------------------------------------------------------------------------------
 
-def remove_files(regexp):
-    grem(".", regexp)
-    raise Exception("not implemented")
+def remove_files(title):
+    grem(".", title + "_[1234567890]{2}\.txt")
+    grem(".", "solve_" + title + "\.m")
+    grem(".", "psat_" + title + "\.m")
+    
+def make_matlab_script(title, simtype):
 
-def make_matlab_script(title):
-    raise Exception("not implemented")
+    if simtype == "powerflow":
+        base_file = "pf_solve.m"
+    elif simtype == "optimalpowerflow":
+        base_file = "opf_solve.m"
 
-def write_contingency(contingency, psat, outfile):
-    raise Exception("not implemented")
+    new_text = open(base_file).read().replace('psatfilename', "psat_" + title)
+    open("solve_" + title + ".m","w").write(new_text)
+    #todo: do i need to close these files? 
+
+def write_contingency(contingency, psat):
+
+    newpsat = copy.deepcopy(psat)
+
+    for kill in contingency.kill["bus"]:
+        newpsat.remove_bus(kill)
+    for kill in contingency.kill["line"]:
+        newpsat.remove_line(kill[0], kill[1])
+    for kill in contingency.kill["generator"]:
+        newpsat.remove_generator(kill)
+
+    if not(len(contingency.supply) == 0 and len(contingency.demand) == 0):
+        raise Exception("not implemented")
+    
+    newpsat.write(open("psat_" + contingency.title + ".m","w"))
 
 def simulate(title):
-    raise Exception("not implemented")
+    proc = subprocess.Popen('matlab -nodesktop -nodisplay -nojvm -nosplash -r solve_' + title,
+                            shell=True,
+                            stdout=subprocess.PIPE,
+                            stderr=subprocess.PIPE,
+                            stdin=subprocess.PIPE)
+    so, se = proc.communicate()
 
-def parselog(title):
-    return SimInLimits(open(title + "_01.txt"))
+#     print "\n\nSE\n\n"
+#     print se 
+
+#     print "\n\nSO\n\n"
+#     print so 
+
+def parselog(title, simtype):
+    return SimInLimits(open("psat_" + title + "_01.txt"), simtype)
  
 #------------------------------------------------------------------------------
 # "main2" function:
@@ -89,12 +135,13 @@ def main2(psat_file, batch_file, outfile):
     contingencies = read_contingency(batch_file)
 
     for contingency in contingencies:
-        remove_files(contingency.title + "_.*\.txt")
-        make_matlab_script(contingency.title)
-        write_contingency(contingency, psat, open("test.m","w"))
+        remove_files(contingency.title)
+        make_matlab_script(contingency.title, contingency.simtype)
+        write_contingency(contingency, psat)
         simulate(contingency.title)
-        contingency.result = parselog(contingency.title)
+        contingency.result = parselog(contingency.title, contingency.simtype)
         print contingency.result
+
     
  
 #------------------------------------------------------------------------------
@@ -103,7 +150,7 @@ def main2(psat_file, batch_file, outfile):
  
 def main():
     """ Parses the command line and call with the correct data.
-"""
+    """
     parser = optparse.OptionParser("usage: program [options] psat_file batch_file")
  
     parser.add_option("-o", "--output", dest="output", metavar="FILE",
@@ -137,17 +184,22 @@ def main():
  
     # Input.
     if len(args) != 2:
+        print "Error: expected 2 arguments got", len(args)
         parser.print_help()
         sys.exit(1)
     else:
         psat_file = open(args[0])
         batch_file = open(args[1])
  
+    t0 = time.time()
     logger.info("Processing %s with %s" % (args[0], args[1]))
     logger.info("===================")
     main2(psat_file, batch_file, outfile)
     logger.info("===================")
  
+    elapsed = time.time() - t0
+    logger.info("Completed in %.3fs." % elapsed)
+
 if __name__ == "__main__":
     main()
